+++ native/examples/2_encoders.cpp
@@ -27,130 +27,6 @@ In these examples we will discuss other ways of laying out data into plaintext
 elements (encoding) that allow more computations without data type overflow, and
 can allow the full plaintext polynomial to be utilized.
 */
-void example_integer_encoder()
-{
-    print_example_banner("Example: Encoders / Integer Encoder");
-
-    /*
-    [IntegerEncoder] (For BFV scheme only)
-
-    The IntegerEncoder encodes integers to BFV plaintext polynomials as follows.
-    First, a binary expansion of the integer is computed. Next, a polynomial is
-    created with the bits as coefficients. For example, the integer
-
-        26 = 2^4 + 2^3 + 2^1
-
-    is encoded as the polynomial 1x^4 + 1x^3 + 1x^1. Conversely, plaintext
-    polynomials are decoded by evaluating them at x=2. For negative numbers the
-    IntegerEncoder simply stores all coefficients as either 0 or -1, where -1 is
-    represented by the unsigned integer plain_modulus - 1 in memory.
-
-    Since encrypted computations operate on the polynomials rather than on the
-    encoded integers themselves, the polynomial coefficients will grow in the
-    course of such computations. For example, computing the sum of the encrypted
-    encoded integer 26 with itself will result in an encrypted polynomial with
-    larger coefficients: 2x^4 + 2x^3 + 2x^1. Squaring the encrypted encoded
-    integer 26 results also in increased coefficients due to cross-terms, namely,
-
-        (1x^4 + 1x^3 + 1x^1)^2 = 1x^8 + 2x^7 + 1x^6 + 2x^5 + 2x^4 + 1x^2;
-
-    further computations will quickly increase the coefficients much more.
-    Decoding will still work correctly in this case (evaluating the polynomial
-    at x=2), but since the coefficients of plaintext polynomials are really
-    integers modulo plain_modulus, implicit reduction modulo plain_modulus may
-    yield unexpected results. For example, adding 1x^4 + 1x^3 + 1x^1 to itself
-    plain_modulus many times will result in the constant polynomial 0, which is
-    clearly not equal to 26 * plain_modulus. It can be difficult to predict when
-    such overflow will take place especially when computing several sequential
-    multiplications.
-
-    The IntegerEncoder is easy to understand and use for simple computations,
-    and can be a good tool to experiment with for users new to Microsoft SEAL.
-    However, advanced users will probably prefer more efficient approaches,
-    such as the BatchEncoder or the CKKSEncoder.
-    */
-    EncryptionParameters parms(scheme_type::bfv);
-    size_t poly_modulus_degree = 4096;
-    parms.set_poly_modulus_degree(poly_modulus_degree);
-    parms.set_coeff_modulus(CoeffModulus::BFVDefault(poly_modulus_degree));
-
-    /*
-    There is no hidden logic behind our choice of the plain_modulus. The only
-    thing that matters is that the plaintext polynomial coefficients will not
-    exceed this value at any point during our computation; otherwise the result
-    will be incorrect.
-    */
-    parms.set_plain_modulus(512);
-    SEALContext context(parms);
-    print_parameters(context);
-    cout << endl;
-
-    KeyGenerator keygen(context);
-    PublicKey public_key = keygen.public_key();
-    SecretKey secret_key = keygen.secret_key();
-    Encryptor encryptor(context, public_key);
-    Evaluator evaluator(context);
-    Decryptor decryptor(context, secret_key);
-
-    /*
-    We create an IntegerEncoder.
-    */
-    IntegerEncoder encoder(context);
-
-    /*
-    First, we encode two integers as plaintext polynomials. Note that encoding
-    is not encryption: at this point nothing is encrypted.
-    */
-    int value1 = 5;
-    Plaintext plain1 = encoder.encode(value1);
-    print_line(__LINE__);
-    cout << "Encode " << value1 << " as polynomial " << plain1.to_string() << " (plain1)," << endl;
-
-    int value2 = -7;
-    Plaintext plain2 = encoder.encode(value2);
-    cout << string(13, ' ') << "encode " << value2 << " as polynomial " << plain2.to_string() << " (plain2)." << endl;
-
-    /*
-    Now we can encrypt the plaintext polynomials.
-    */
-    Ciphertext encrypted1, encrypted2;
-    print_line(__LINE__);
-    cout << "Encrypt plain1 to encrypted1 and plain2 to encrypted2." << endl;
-    encryptor.encrypt(plain1, encrypted1);
-    encryptor.encrypt(plain2, encrypted2);
-    cout << "    + Noise budget in encrypted1: " << decryptor.invariant_noise_budget(encrypted1) << " bits" << endl;
-    cout << "    + Noise budget in encrypted2: " << decryptor.invariant_noise_budget(encrypted2) << " bits" << endl;
-
-    /*
-    As a simple example, we compute (-encrypted1 + encrypted2) * encrypted2.
-    */
-    Ciphertext encrypted_result;
-    print_line(__LINE__);
-    cout << "Compute encrypted_result = (-encrypted1 + encrypted2) * encrypted2." << endl;
-    evaluator.negate(encrypted1, encrypted_result);
-    evaluator.add_inplace(encrypted_result, encrypted2);
-    evaluator.multiply_inplace(encrypted_result, encrypted2);
-    cout << "    + Noise budget in encrypted_result: " << decryptor.invariant_noise_budget(encrypted_result) << " bits"
-         << endl;
-    Plaintext plain_result;
-    print_line(__LINE__);
-    cout << "Decrypt encrypted_result to plain_result." << endl;
-    decryptor.decrypt(encrypted_result, plain_result);
-
-    /*
-    Print the result plaintext polynomial. The coefficients are not even close
-    to exceeding our plain_modulus, 512.
-    */
-    cout << "    + Plaintext polynomial: " << plain_result.to_string() << endl;
-
-    /*
-    Decode to obtain an integer result.
-    */
-    print_line(__LINE__);
-    cout << "Decode plain_result." << endl;
-    cout << "    + Decoded integer: " << encoder.decode_int32(plain_result);
-    cout << "...... Correct." << endl;
-}
 
 void example_batch_encoder()
 {
@@ -166,7 +42,7 @@ void example_batch_encoder()
     several orders of magnitude in fully vectorizable computations. Thus, in all
     but the simplest computations, batching should be the preferred method to use
     with BFV, and when used properly will result in implementations outperforming
-    anything done with the IntegerEncoder.
+    anything done without batching.
     */
     EncryptionParameters parms(scheme_type::bfv);
     size_t poly_modulus_degree = 8192;
@@ -369,7 +245,7 @@ void example_ckks_encoder()
 
     /*
     To create CKKS plaintexts we need a special encoder: there is no other way
-    to create them. The IntegerEncoder and BatchEncoder cannot be used with the
+    to create them. The BatchEncoder cannot be used with the
     CKKS scheme. The CKKSEncoder encodes vectors of real or complex numbers into
     Plaintext objects, which can subsequently be encrypted. At a high level this
     looks a lot like what BatchEncoder does for the BFV scheme, but the theory
@@ -468,7 +344,6 @@ void example_encoders()
     /*
     Run all encoder examples.
     */
-    example_integer_encoder();
     example_batch_encoder();
     example_ckks_encoder();
 }